Move `ScreencastSource` to `fht-compositor-ipc`
